requirement.txt
# main libraries and dependencies
keras==2.1.5
# tensorflow==1.4.1
tensorflow-gpu==1.4
pandas
h5py==2.8.0
nilearn==0.5.0
numpy==1.15.4
scikit-learn
matplotlib
opencv-python
tqdm